feat: #3 authViewModel 적용
