+++ moyeo/moyeo/View/LoginView.swift
@@ -2,7 +2,7 @@
 //  LoginView.swift
 //  moyeo
 //
-//  Created by Woowon Kang on 7/25/24.
+//  Created by Chang Jonghyeon on 7/18/24.
 //
 
 import SwiftUI
@@ -10,28 +10,44 @@ import AuthenticationServices
 
 struct LoginView: View {
     @EnvironmentObject var appViewModel: AppViewModel
-
+    
     var body: some View {
         VStack {
             Text("모여 로고")
                 .font(.title)
             
-            Button(action: {
-                // 임의로 로그인 성공 처리
-                appViewModel.handleSignInWithApple()
-            }) {
-                Text("임시 Apple Login Button")
-                    .foregroundColor(.white)
-                    .padding()
-                    
-            }
-            .frame(width: 360, height: 50)
-            .background(Color.black)
-            .cornerRadius(14)
+            AppleSigninButton()
+
         }
     }
 }
 
+struct AppleSigninButton: View {
+    @EnvironmentObject var authViewModel: AuthViewModel
+    
+    var body: some View {
+        
+        SignInWithAppleButton(
+            .signIn,
+            onRequest: { request in
+                Task {
+                    await authViewModel.appleLogin(request: request)
+                }
+            },
+            onCompletion: { result in
+                Task {
+                    await authViewModel.appleLoginCompletion(result: result)
+                }
+            }
+        )
+        .signInWithAppleButtonStyle(.black)
+        .frame(width: 361, height: 50)
+        
+    }
+    
+    
+    
+}
 
 #Preview {
     LoginView()
